requirements.txt
GitPython
docling